fix configuration, remove debug code, optimize nbtfactory

+++ src/main/java/noitemabuse/util/FileLogger.java
@@ -12,13 +12,13 @@ import java.util.Date;
 public class FileLogger {
     private SimpleDateFormat format = new SimpleDateFormat("[yyyy-MM-dd HH:mm:ss] ");
     private Date date = new Date();
-    private final PrintWriter writer;
+    public final PrintWriter writer;
     private Thread thread = new Thread(new Flusher());
-    private final Object lock = new Object();
+    public static final Object lock = new Object();
+    public boolean requireFlush = false;
 
     public FileLogger(File file) throws IOException {
-        file.getParentFile().mkdirs();
-        file.createNewFile();
+        if (!file.exists() && !(file.getParentFile() != null && file.getParentFile().mkdirs()) && !file.createNewFile()) throw new IOException("Could not create log file");
         writer = new PrintWriter(new FileWriter(file, true));
         thread.start();
     }
@@ -32,6 +32,10 @@ public class FileLogger {
         thread.interrupt();
     }
 
+    public PrintWriter getWriter() {
+        return writer;
+    }
+
     public void log(String string) {
         date.setTime(System.currentTimeMillis());
         String timestamp = format.format(date);
@@ -41,6 +45,7 @@ public class FileLogger {
     private void println(String str) {
         writer.println(str);
         synchronized (lock) {
+            requireFlush = true;
             lock.notifyAll();
         }
     }
@@ -51,9 +56,12 @@ public class FileLogger {
             try {
                 while (true) {
                     synchronized (lock) {
-                        lock.wait();
+                        while (!requireFlush) {
+                            lock.wait();
+                        }
+                        writer.flush();
+                        requireFlush = false;
                     }
-                    writer.flush();
                 }
             } catch (InterruptedException ex) {} finally {
                 writer.close();
